+++ packages/terra-popup/src/terra-dev-site/test/popup/ResizedContentPopup.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import classNames from 'classnames/bind';
 import Popup from '../../../Popup';
 import TestPopupContent from '../common/TestPopupContent';
-import styles from './AutomaticHeightAndResizedContentCommon.scss';
+import styles from './AutomaticHeightAndResizedContentCommon.module.scss';
 
 const cx = classNames.bind(styles);