+++ packages/terra-date-picker/CHANGELOG.md
@@ -8,6 +8,7 @@ Unreleased
 
 ### Changed
 * Removed extra inline styles.
+* Renamed test and example SCSS files to end with `.module.scss` extension
 
 4.4.0 - (June 19, 2019)
 ------------------